Move notifications to standalone directory
