+++ plugin-hrm-form/src/HrmFormPlugin.tsx
@@ -40,7 +40,8 @@ import {
 import { setUpSharedStateClient } from './utils/sharedState';
 import { FeatureFlags } from './types/types';
 import { setUpReferrableResources } from './components/resources/setUpReferrableResources';
-import { subscribeNewMessageAlertOnPluginInit, subscribeReservedTaskAlert } from './utils/audioNotifications';
+import { subscribeNewMessageAlertOnPluginInit } from './notifications/newMessage';
+import { subscribeReservedTaskAlert } from './notifications/reservedTask';
 import { setUpCounselorToolkits } from './components/toolkits/setUpCounselorToolkits';
 
 const PLUGIN_NAME = 'HrmFormPlugin';
@@ -115,6 +116,7 @@ const setUpComponents = (
   if (maskIdentifiers) {
     // Mask the identifiers in all default channels
     Channels.maskIdentifiersForDefaultChannels();
+
     // Mask the username within the messable bubbles in an conversation
     Flex.MessagingCanvas.defaultProps.memberDisplayOptions = {
       theirDefaultName: 'XXXXXX',
